replaced CachedVerifier object by functions and global map
